backend/requirements.txt
Django>=4.0
djangorestframework
gunicorn
django-redis
python-decouple
django-cors-headers
Pillow